// src/app/generated/models/DataCorrectnessConfirmationDetailDtoOut.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mild Blue - Covid Vaxx
 * Covid Vaxx API
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: support@mild.blue
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PersonnelDtoOut,
    PersonnelDtoOutFromJSON,
    PersonnelDtoOutFromJSONTyped,
    PersonnelDtoOutToJSON,
} from './';

/**
 * 
 * @export
 * @interface DataCorrectnessConfirmationDetailDtoOut
 */
export interface DataCorrectnessConfirmationDetailDtoOut {
    /**
     * 
     * @type {Date}
     * @memberof DataCorrectnessConfirmationDetailDtoOut
     */
    checked: Date;
    /**
     * 
     * @type {boolean}
     * @memberof DataCorrectnessConfirmationDetailDtoOut
     */
    dataAreCorrect: boolean;
    /**
     * 
     * @type {PersonnelDtoOut}
     * @memberof DataCorrectnessConfirmationDetailDtoOut
     */
    doctor: PersonnelDtoOut;
    /**
     * 
     * @type {string}
     * @memberof DataCorrectnessConfirmationDetailDtoOut
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof DataCorrectnessConfirmationDetailDtoOut
     */
    notes?: string | null;
    /**
     * 
     * @type {PersonnelDtoOut}
     * @memberof DataCorrectnessConfirmationDetailDtoOut
     */
    nurse?: PersonnelDtoOut;
}

export function DataCorrectnessConfirmationDetailDtoOutFromJSON(json: any): DataCorrectnessConfirmationDetailDtoOut {
    return DataCorrectnessConfirmationDetailDtoOutFromJSONTyped(json, false);
}

export function DataCorrectnessConfirmationDetailDtoOutFromJSONTyped(json: any, ignoreDiscriminator: boolean): DataCorrectnessConfirmationDetailDtoOut {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'checked': (new Date(json['checked'])),
        'dataAreCorrect': json['dataAreCorrect'],
        'doctor': PersonnelDtoOutFromJSON(json['doctor']),
        'id': json['id'],
        'notes': !exists(json, 'notes') ? undefined : json['notes'],
        'nurse': !exists(json, 'nurse') ? undefined : PersonnelDtoOutFromJSON(json['nurse']),
    };
}

export function DataCorrectnessConfirmationDetailDtoOutToJSON(value?: DataCorrectnessConfirmationDetailDtoOut | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'checked': (value.checked.toISOString()),
        'dataAreCorrect': value.dataAreCorrect,
        'doctor': PersonnelDtoOutToJSON(value.doctor),
        'id': value.id,
        'notes': value.notes,
        'nurse': PersonnelDtoOutToJSON(value.nurse),
    };
}